requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmdstanpy==0.9.5
convertdate==2.2.2
cycler==0.10.0
Cython==0.29.21
ephem==3.7.7.1
fbprophet==0.7.1
Flask==1.1.2
gunicorn==20.0.4
holidays==0.10.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.2
Pillow==7.2.0
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
python-firebase==1.2
pytz==2020.1
requests==2.24.0
setuptools-git==1.2
six==1.15.0
tqdm==4.48.2
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
